% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-docs.R
\name{add_all_left}
\alias{add_all_left}
\title{Add multiple elements to the left}
\usage{
add_all_left(...)
}
\arguments{
\item{t}{FingerTree.}

\item{els}{Elements to add.}
}
\description{
Add multiple elements to the left
}
