% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-docs.R
\name{add_all_right}
\alias{add_all_right}
\title{Add multiple elements to the right}
\usage{
add_all_right(...)
}
\arguments{
\item{t}{FingerTree.}

\item{els}{Elements to add.}
}
\description{
Add multiple elements to the right
}
