% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-docs.R
\name{app3}
\alias{app3}
\title{Generalized concatenation helper}
\usage{
app3(...)
}
\arguments{
\item{xs}{Left tree.}

\item{ts}{List of elements to insert between trees.}

\item{ys}{Right tree.}
}
\description{
Generalized concatenation helper
}
