% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append.flexseq}
\alias{append.flexseq}
\title{Append an element to a flexseq}
\usage{
\method{append}{flexseq}(x, values, ...)
}
\arguments{
\item{x}{A `flexseq`.}

\item{values}{Element to append.}

\item{...}{Unused.}
}
\value{
Updated `flexseq`.
}
\description{
Append an element to a flexseq
}
\examples{
x <- as_flexseq(letters[1:4])
x

x2 <- append(x, "p")
x2

# flexseqs can also hold nested or mixed types
x3 <- append(x2, c(8, 9))
x3

# appending to a named sequence requires a named element
x4 <- as_flexseq(list(one = 1, two = 2, three = 3))
new_el <- 4
names(new_el) <- "four"
x5 <- append(x4, new_el)
x5
}
