% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append}
\alias{append}
\title{Append an element}
\usage{
append(t, x)
}
\arguments{
\item{t}{FingerTree.}

\item{x}{Element to append.}
}
\value{
Updated tree.
  If `t` is name-indexed, appending unnamed elements is invalid.
}
\description{
Append an element
}
\examples{
t <- as_flexseq(letters[1:3])
t2 <- append(t, "d")
t2[[4]]

# Compose with prepend and reduce
cat_m <- measure_monoid(paste0, "", as.character)
fold_right(prepend(t2, "z"), cat_m)

# Append to a named tree with an explicit element name
tn <- as_flexseq(setNames(as.list(1:2), c("a", "b")))
tn2 <- append(tn, stats::setNames(3, "c"))
tn2[["c"]]
}
