% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-methods.R
\name{as.list.interval_index}
\alias{as.list.interval_index}
\title{Coerce Interval Index to List}
\usage{
\method{as.list}{interval_index}(x, ...)
}
\arguments{
\item{x}{An `interval_index`.}

\item{...}{Unused.}
}
\value{
A plain list of payload elements in interval order.
}
\description{
Coerce Interval Index to List
}
\details{
This returns payload values only.
}
\examples{
ix <- interval_index("a", "b", "c", start = c(3, 1, 2), end = c(4, 2, 3))
as.list(ix)
}
