% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_multiset-core.R
\name{as_ordered_multiset}
\alias{as_ordered_multiset}
\title{Build an Ordered Multiset from elements}
\usage{
as_ordered_multiset(x, keys = NULL, monoids = NULL)
}
\arguments{
\item{x}{Elements to add.}

\item{keys}{Scalar key values matching `x` length.}

\item{monoids}{Optional additional named list of `measure_monoid` objects.}
}
\value{
An `ordered_multiset`.
}
\description{
Build an Ordered Multiset from elements
}
\examples{
ms <- as_ordered_multiset(c(4, 1, 2, 1), keys = c(4, 1, 2, 1))
ms
count_key(ms, 1)
}
