% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dollar.R
\name{$.flexseq}
\alias{$.flexseq}
\title{Extract a named element with `$`}
\usage{
\method{$}{flexseq}(x, name)
}
\arguments{
\item{x}{A `flexseq`.}

\item{name}{Element name.}
}
\value{
The matched element.
}
\description{
Exact-name lookup on named trees, equivalent to `x[[\"name\"]]`.
}
\examples{
x <- as_flexseq(setNames(as.list(1:3), c("a", "b", "c")))
x$b
}
