% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dollar.R
\name{$<-.flexseq}
\alias{$<-.flexseq}
\title{Replace a named element with `$<-`}
\usage{
\method{$}{flexseq}(x, name) <- value
}
\arguments{
\item{x}{A `flexseq`.}

\item{name}{Element name.}

\item{value}{Replacement element.}
}
\value{
Updated tree.
}
\description{
Exact-name replacement on named trees, equivalent to `x[[\"name\"]] <- value`.
}
\examples{
t <- as_flexseq(setNames(as.list(1:3), c("a", "b", "c")))
t$b <- 20
t$b
}
