% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_multiset-core.R
\name{delete_one}
\alias{delete_one}
\title{Delete one element by key}
\usage{
delete_one(x, key_value)
}
\arguments{
\item{x}{An `ordered_multiset`.}

\item{key_value}{Key value to delete.}
}
\value{
Updated `ordered_multiset`.
}
\description{
Delete one element by key
}
