% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_multiset-core.R
\name{difference_ms}
\alias{difference_ms}
\title{Multiset difference (bag semantics)}
\usage{
difference_ms(x, y)
}
\arguments{
\item{x}{Left `ordered_multiset`.}

\item{y}{Right `ordered_multiset`.}
}
\value{
`ordered_multiset` where multiplicity is `pmax(count_x - count_y, 0)`.
}
\description{
Multiset difference (bag semantics)
}
