% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-empty-tree.R
\name{empty_tree}
\alias{empty_tree}
\title{Create an Empty Structural Tree}
\usage{
empty_tree(monoids = NULL)
}
\arguments{
\item{monoids}{Optional named list of \code{measure_monoid} objects.}
}
\value{
An empty finger tree with structural \code{monoids} and \code{measures} attrs.
}
\description{
Create an Empty Structural Tree
}
\keyword{internal}
