% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-core.R
\name{extract_min}
\alias{extract_min}
\title{Extract minimum-priority element}
\usage{
extract_min(q)
}
\arguments{
\item{q}{A `priority_queue`.}
}
\value{
List with `element`, `priority`, and updated `queue`.
}
\description{
Extract minimum-priority element
}
\examples{
x <- priority_queue("a", "b", "c", priorities = c(2, 1, 1))
out <- extract_min(x)
out$element
out$priority
out$queue
}
