% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-interval_index-core.R
\name{find_point}
\alias{find_point}
\title{Find intervals containing a point}
\usage{
find_point(x, point, bounds = NULL)
}
\arguments{
\item{x}{An `interval_index`.}

\item{point}{Query point.}

\item{bounds}{Optional boundary override. One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}
}
\value{
An `interval_index` slice of matches.
}
\description{
Find intervals containing a point
}
