% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-interval_index-core.R
\name{find_within}
\alias{find_within}
\title{Find intervals within a query interval}
\usage{
find_within(x, start, end, bounds = NULL)
}
\arguments{
\item{x}{An `interval_index`.}

\item{start}{Query interval start.}

\item{end}{Query interval end.}

\item{bounds}{Optional boundary override. One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}
}
\value{
An `interval_index` slice of matches.
}
\description{
Find intervals within a query interval
}
