% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingertree-package.R
\docType{package}
\name{fingertree-package}
\alias{fingertree}
\alias{fingertree-package}
\title{fingertree}
\description{
Monoid-annotated 2-3 finger trees.
}
\author{
\strong{Maintainer}: Shawn T. O'Neil \email{shawn@tislab.org}

}
