% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexseq.R
\name{fold_left}
\alias{fold_left}
\title{Fold Left Over a Sequence}
\usage{
fold_left(x, monoid)
}
\arguments{
\item{x}{A `flexseq`.}

\item{monoid}{A `measure_monoid` specification.}
}
\value{
Reduced value.
}
\description{
Fold Left Over a Sequence
}
