% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_multiset-core.R
\name{insert_ms}
\alias{insert_ms}
\title{Insert an element into an ordered multiset}
\usage{
insert_ms(x, element)
}
\arguments{
\item{x}{An `ordered_multiset`.}

\item{element}{Element to insert.}
}
\value{
Updated `ordered_multiset`.
}
\description{
Insert an element into an ordered multiset
}
