% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_multiset-core.R
\name{intersection_ms}
\alias{intersection_ms}
\title{Multiset intersection (bag semantics)}
\usage{
intersection_ms(x, y)
}
\arguments{
\item{x}{Left `ordered_multiset`.}

\item{y}{Right `ordered_multiset`.}
}
\value{
`ordered_multiset` where multiplicity is `min(count_x, count_y)`.
}
\description{
Multiset intersection (bag semantics)
}
