% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priority_queue.R
\name{is_empty}
\alias{is_empty}
\title{Check whether a priority queue is empty}
\usage{
is_empty(q)
}
\arguments{
\item{q}{A `priority_queue`.}
}
\value{
Logical scalar.
}
\description{
Check whether a priority queue is empty
}
\examples{
x <- priority_queue()
is_empty(x)

x2 <- priority_queue("a", priorities = 1)
is_empty(x2)
}
