% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-ordered_sequence-methods.R
\name{length.ordered_sequence}
\alias{length.ordered_sequence}
\title{Ordered Sequence Length}
\usage{
\method{length}{ordered_sequence}(x)
}
\arguments{
\item{x}{An \code{ordered_sequence}.}
}
\value{
Integer length.
}
\description{
Ordered Sequence Length
}
\details{
Uses cached size metadata and runs in O(1).
}
\examples{
x <- ordered_sequence("a", "b", keys = c(2, 1))
length(x)

length(ordered_sequence())
}
