% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-methods.R
\name{length.priority_queue}
\alias{length.priority_queue}
\title{Priority Queue Length}
\usage{
\method{length}{priority_queue}(x)
}
\arguments{
\item{x}{A \code{priority_queue}.}
}
\value{
Integer length.
}
\description{
Priority Queue Length
}
\details{
Uses cached size metadata and runs in O(1).
}
\examples{
q <- priority_queue("a", "b", priorities = c(2, 1))
length(q)

length(priority_queue())
}
