% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-docs.R
\name{MeasuredReducer}
\alias{MeasuredReducer}
\title{Measured reducer}
\usage{
MeasuredReducer(...)
}
\arguments{
\item{f}{Reduction function.}

\item{i}{Identity element.}

\item{measure}{Function mapping an element to a measure.}
}
\description{
Measured reducer
}
