% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureMonoid.R
\name{MeasureMonoid}
\alias{MeasureMonoid}
\title{Construct a MeasureMonoid}
\usage{
MeasureMonoid(f, i, measure)
}
\arguments{
\item{f}{Associative binary function over measure values.}

\item{i}{Identity measure value for `f`.}

\item{measure}{Function mapping a raw element to a measure value.}
}
\value{
An object of class `MeasureMonoid`.
}
\description{
Construct a MeasureMonoid
}
