% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-docs.R
\name{Node3}
\alias{Node3}
\title{Node with three elements}
\usage{
Node3(...)
}
\arguments{
\item{x}{First element.}

\item{y}{Second element.}

\item{z}{Third element.}
}
\description{
Node with three elements
}
