% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_multiset-core.R
\name{ordered_multiset}
\alias{ordered_multiset}
\title{Construct an Ordered Multiset}
\usage{
ordered_multiset(..., key = identity, monoids = NULL)
}
\arguments{
\item{...}{Elements to add.}

\item{key}{Function extracting a scalar key from each element.}

\item{monoids}{Optional additional named list of `measure_monoid` objects.}
}
\value{
An `ordered_multiset`.
}
\description{
Construct an Ordered Multiset
}
\examples{
ms <- ordered_multiset("bb", "a", "ccc", key = nchar)
ms
lower_bound(ms, 2)
}
