% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-query-api.R
\name{peek_all_containing}
\alias{peek_all_containing}
\title{Peek All Intervals Containing a Query Interval}
\usage{
peek_all_containing(x, start, end, bounds = NULL)
}
\arguments{
\item{x}{An `interval_index`.}

\item{start}{Query interval start.}

\item{end}{Query interval end.}

\item{bounds}{Optional boundary override. One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}
}
\value{
An `interval_index` slice of all matches (possibly empty).
}
\description{
Peek All Intervals Containing a Query Interval
}
\details{
The returned `interval_index` can be inspected with [as.list()].
}
\examples{
ix <- interval_index("a", "b", "c", start = c(1, 2, 4), end = c(6, 5, 7))
as.list(peek_all_containing(ix, 2, 4))
}
