% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-queue-ops.R
\name{peek_all_max}
\alias{peek_all_max}
\title{Peek all maximum-priority elements}
\usage{
peek_all_max(x)
}
\arguments{
\item{x}{A `priority_queue`.}
}
\value{
A `priority_queue` containing all maximum-priority elements in stable
  FIFO order.
}
\description{
Peek all maximum-priority elements
}
