% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-ends.R
\name{peek_back}
\alias{peek_back}
\title{Peek at the Back Element}
\usage{
peek_back(x)
}
\arguments{
\item{x}{A `flexseq`.}
}
\value{
Last element, or `NULL` when `x` is empty.
}
\description{
Returns the last element without modifying the sequence.
}
\examples{
x <- flexseq("a", "b", "c")
peek_back(x)

peek_back(flexseq())
}
