% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-ends.R
\name{peek_front}
\alias{peek_front}
\title{Peek at the Front Element}
\usage{
peek_front(x)
}
\arguments{
\item{x}{A `flexseq`.}
}
\value{
First element, or `NULL` when `x` is empty.
}
\description{
Returns the first element without modifying the sequence.
}
\examples{
x <- flexseq("a", "b", "c")
peek_front(x)

peek_front(flexseq())
}
