% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-queue-ops.R
\name{peek_max}
\alias{peek_max}
\title{Peek maximum-priority element}
\usage{
peek_max(x)
}
\arguments{
\item{x}{A `priority_queue`.}
}
\value{
Element with maximum priority (stable on ties), or `NULL` when empty.
}
\description{
Peek maximum-priority element
}
\examples{
x <- priority_queue("a", "b", "c", priorities = c(2, 3, 3))
x
peek_max(x)
}
