% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-methods.R
\name{plot.interval_index}
\alias{plot.interval_index}
\title{Plot an Interval Index Tree}
\usage{
\method{plot}{interval_index}(x, ...)
}
\arguments{
\item{x}{An `interval_index`.}

\item{...}{Passed to the internal tree plotting routine.}
}
\description{
Plots the underlying tree structure used by an `interval_index`.
}
\examples{
ix <- interval_index("a", "b", "c", start = c(1, 3, 5), end = c(2, 4, 6))
plot(ix)
}
