% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-ordered_sequence-methods.R
\name{plot.ordered_sequence}
\alias{plot.ordered_sequence}
\title{Plot an Ordered Sequence Tree}
\usage{
\method{plot}{ordered_sequence}(x, ...)
}
\arguments{
\item{x}{An `ordered_sequence`.}

\item{...}{Passed to the internal tree plotting routine.}
}
\description{
Plots the underlying tree structure used by an `ordered_sequence`.
}
\examples{
x <- ordered_sequence("a", "b", "c", keys = c(2, 1, 3))
plot(x)
}
