% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-query-api.R
\name{pop_all_within}
\alias{pop_all_within}
\title{Pop All Intervals Within a Query Interval}
\usage{
pop_all_within(x, start, end, bounds = NULL)
}
\arguments{
\item{x}{An `interval_index`.}

\item{start}{Query interval start.}

\item{end}{Query interval end.}

\item{bounds}{Optional boundary override. One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}
}
\value{
A list with `elements` and `remaining`, both `interval_index` objects.
}
\description{
Pop All Intervals Within a Query Interval
}
\details{
Use [as.list()] to convert `elements` to a standard R list.
}
\examples{
ix <- interval_index("a", "b", "c", start = c(1, 2, 4), end = c(6, 3, 5))
out <- pop_all_within(ix, 1, 4)
as.list(out$elements)
}
