% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepend.R
\name{prepend}
\alias{prepend}
\title{Prepend an element}
\usage{
prepend(t, x)
}
\arguments{
\item{t}{FingerTree.}

\item{x}{Element to prepend.}
}
\value{
Updated tree.
  If `t` is name-indexed, prepending unnamed elements is invalid.
}
\description{
Prepend an element
}
\examples{
t <- tree_from(letters[2:4])
t2 <- prepend(t, "a")
t2[[1]]

# Compose with append and reduce
cat_m <- MeasureMonoid(paste0, "", as.character)
reduce_left(append(t2, "z"), cat_m)

# Prepend to a named tree with an explicit element name
tn <- tree_from(setNames(as.list(2:3), c("b", "c")))
tn2 <- prepend(tn, stats::setNames(1, "a"))
tn2[["a"]]
}
