% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_right.R
\name{reduce_right}
\alias{reduce_right}
\title{Reduce from the right}
\usage{
reduce_right(t, monoid)
}
\arguments{
\item{t}{FingerTree.}

\item{monoid}{A `MeasureMonoid` object.}
}
\value{
Reduced value.
}
\description{
Reduce from the right
}
\examples{
t <- as_flexseq(1:5)
sum_m <- measure_monoid(`+`, 0, as.numeric)
reduce_right(t, sum_m)

cat_m <- measure_monoid(paste0, "", as.character)
reduce_right(as_flexseq(letters[1:4]), cat_m)
}
