% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{split}
\alias{split}
\title{Split tree into left and right parts}
\usage{
split(t, predicate, monoid = NULL)
}
\arguments{
\item{t}{FingerTree.}

\item{predicate}{Function on measure values.}

\item{monoid}{Optional monoid; falls back to tree attribute.}
}
\value{
A list with `left` and `right`.
}
\description{
Split tree into left and right parts
}
