% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-str.R
\name{str.flexseq}
\alias{str.flexseq}
\title{Display Internal Structure of a flexseq}
\usage{
\method{str}{flexseq}(object, ...)
}
\arguments{
\item{object}{A \code{flexseq}.}

\item{...}{Passed to \code{\link[utils:str]{utils::str()}}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Display Internal Structure of a flexseq
}
\examples{
x <- flexseq(a = 1, b = list(k = 2))
str(x)
}
\keyword{internal}
