% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{[.FingerTree}
\alias{[.FingerTree}
\title{Subset a finger tree by positive integer indices}
\usage{
\method{[}{FingerTree}(x, i, ...)
}
\arguments{
\item{x}{FingerTree.}

\item{i}{Positive integer index vector.}

\item{...}{Unused.}
}
\value{
A new FingerTree containing selected elements in index order.
}
\description{
Subset a finger tree by positive integer indices
}
