% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-indexing.R
\name{[[<-.flexseq}
\alias{[[<-.flexseq}
\title{Replace one element by position or unique name}
\usage{
\method{[[}{flexseq}(x, i) <- value
}
\arguments{
\item{x}{A `flexseq`.}

\item{i}{Positive scalar integer index, or scalar character element name.}

\item{value}{Replacement element.}
}
\value{
A new `flexseq` with one element replaced.
}
\description{
Replace one element by position or unique name
}
\examples{
x <- as_flexseq(letters[1:4])
x2 <- x
x2[[2]] <- "ZZ"
x2

x3 <- as_flexseq(setNames(as.list(1:3), c("x", "y", "z")))
x3[["y"]] <- 99
x3

# assigning NULL removes one element
x4 <- as_flexseq(letters[1:4])
x4[[2]] <- NULL
x4
}
