% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{[<-.FingerTree}
\alias{[<-.FingerTree}
\title{Replace selected elements by positive integer indices}
\usage{
\method{[}{FingerTree}(x, i) <- value
}
\arguments{
\item{x}{FingerTree.}

\item{i}{Positive integer index vector.}

\item{value}{Replacement values; must have exactly same length as `i`.}
}
\value{
A new FingerTree with selected elements replaced.
}
\description{
Replace selected elements by positive integer indices
}
