% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_multiset-core.R
\name{union_ms}
\alias{union_ms}
\title{Multiset union (bag semantics)}
\usage{
union_ms(x, y)
}
\arguments{
\item{x}{Left `ordered_multiset`.}

\item{y}{Right `ordered_multiset`.}
}
\value{
`ordered_multiset` where multiplicity is `max(count_x, count_y)`.
}
\description{
Multiset union (bag semantics)
}
