% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-api-generics.R
\name{add_monoids}
\alias{add_monoids}
\title{Add/merge monoids on an existing tree}
\usage{
add_monoids(t, monoids, overwrite = FALSE)
}
\arguments{
\item{t}{FingerTree.}

\item{monoids}{Named list of `measure_monoid` objects to add.}

\item{overwrite}{Logical; whether overlapping names replace existing monoids.}
}
\value{
A persistent copy with recomputed cached measures.
}
\description{
Add/merge monoids on an existing tree
}
\seealso{
[add_monoids.flexseq()], [add_monoids.priority_queue()],
  [add_monoids.ordered_sequence()], [add_monoids.interval_index()]
}
