% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_monoids.R
\name{add_monoids}
\alias{add_monoids}
\title{Add/merge monoids on an existing tree}
\usage{
add_monoids(t, monoids, overwrite = FALSE)
}
\arguments{
\item{t}{FingerTree.}

\item{monoids}{Named list of `measure_monoid` objects to add.}

\item{overwrite}{Logical; whether overlapping names replace existing monoids.}
}
\value{
A persistent copy with recomputed cached measures.
}
\description{
Add/merge monoids on an existing tree
}
\examples{
x <- as_flexseq(1:5)
x

sum_m <- measure_monoid(`+`, 0, as.numeric)
x2 <- add_monoids(x, list(sum = sum_m))
x2
attr(x2, "measures")$sum

# replace an existing monoid definition
sum2 <- measure_monoid(function(a, b) a + 2 * b, 0, as.numeric)
x3 <- add_monoids(x2, list(sum = sum2), overwrite = TRUE)
attr(x3, "measures")$sum
}
