% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append.flexseq}
\alias{append.flexseq}
\title{Append an element to a flexseq}
\usage{
\method{append}{flexseq}(x, values, ...)
}
\arguments{
\item{x}{A `flexseq`.}

\item{values}{Element to append.}

\item{...}{Unused.}
}
\value{
Updated `flexseq`.
}
\description{
Append an element to a flexseq
}
\examples{
t <- as_flexseq(letters[1:3])
t2 <- append(t, "d")
t2[[4]]

# Append to a named sequence
tn <- as_flexseq(setNames(as.list(1:2), c("a", "b")))
tn2 <- append(tn, stats::setNames(3, "c"))
tn2[["c"]]
}
