% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{append}
\alias{append}
\title{Append an element}
\usage{
append(t, x, reducer = NULL)
}
\arguments{
\item{t}{FingerTree.}

\item{x}{Element to append.}

\item{reducer}{Optional reducer to store on the tree.}
}
\value{
Updated tree.
}
\description{
Append an element
}
