% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-core.R
\name{as.list.flexseq}
\alias{as.list.flexseq}
\title{Coerce a Sequence to Base List}
\usage{
\method{as.list}{flexseq}(x, ...)
}
\arguments{
\item{x}{A `flexseq`.}

\item{...}{Unused.}
}
\value{
A base R list of sequence elements.
}
\description{
Returns elements in left-to-right sequence order.
}
