% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-interval_index-core.R
\name{as.list.interval_index}
\alias{as.list.interval_index}
\title{Coerce Interval Index to List}
\usage{
\method{as.list}{interval_index}(x, ...)
}
\arguments{
\item{x}{An `interval_index`.}

\item{...}{Unused.}
}
\value{
A plain list of payload elements in interval order.
}
\description{
Coerce Interval Index to List
}
