% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_sequence-core.R
\name{as.list.ordered_sequence}
\alias{as.list.ordered_sequence}
\title{Coerce Ordered Sequence to List}
\usage{
\method{as.list}{ordered_sequence}(x, ...)
}
\arguments{
\item{x}{An `ordered_sequence`.}

\item{...}{Unused.}
}
\value{
A plain list of elements in key order.
}
\description{
Coerce Ordered Sequence to List
}
