% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-methods.R
\name{as.list.priority_queue}
\alias{as.list.priority_queue}
\title{Coerce Priority Queue to List}
\usage{
\method{as.list}{priority_queue}(x, ...)
}
\arguments{
\item{x}{A `priority_queue`.}

\item{...}{Unused.}
}
\value{
A plain list of queue entry records.
}
\description{
Returns queue entries as a plain list of records with fields `item` and
`priority`, in queue sequence order.
}
\examples{
q <- priority_queue("a", "b", priorities = c(2, 1))
as.list(q)
}
