% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-api-generics.R
\name{as_flexseq}
\alias{as_flexseq}
\title{Coerce to flexseq}
\usage{
as_flexseq(x)
}
\arguments{
\item{x}{Input vector/list-like object.}
}
\value{
A `flexseq` object.
}
\description{
Coerce to flexseq
}
