% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexseq.R
\name{as_flexseq}
\alias{as_flexseq}
\title{Coerce to a Persistent Flexible Sequence}
\usage{
as_flexseq(x, values = NULL, monoids = NULL)
}
\arguments{
\item{x}{Input vector/list-like object.}

\item{values}{Optional parallel values (same length as `x`).}

\item{monoids}{Optional named list of measure monoids.}
}
\value{
A `flexseq` object.
}
\description{
Coerce to a Persistent Flexible Sequence
}
\examples{
as_flexseq(1:5)
as_flexseq(setNames(as.list(letters[1:3]), c("k1", "k2", "k3")))
}
