% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-interval_index-core.R
\name{as_interval_index}
\alias{as_interval_index}
\title{Build an Interval Index from elements and interval bounds}
\usage{
as_interval_index(x, start, end, bounds = "[)")
}
\arguments{
\item{x}{Elements to add.}

\item{start}{Scalar start endpoints (same length as `x`).}

\item{end}{Scalar end endpoints (same length as `x`).}

\item{bounds}{One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}
}
\value{
An `interval_index`.
}
\description{
Build an Interval Index from elements and interval bounds
}
\examples{
ix <- as_interval_index(c("a", "b", "c"), start = c(1, 2, 2), end = c(3, 2, 4))
ix
as.list(peek_point(ix, 2, which = "all"))
}
