% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-core.R
\name{c.flexseq}
\alias{c.flexseq}
\title{Concatenate Sequences}
\usage{
\method{c}{flexseq}(..., recursive = FALSE)
}
\arguments{
\item{...}{`flexseq` objects.}

\item{recursive}{Unused; must be `FALSE`.}
}
\value{
A concatenated `flexseq`.
}
\description{
Concatenate Sequences
}
