% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-concat-trees.R
\name{concat_trees}
\alias{concat_trees}
\title{Concatenate Two Structural Trees}
\usage{
concat_trees(x, y)
}
\arguments{
\item{x}{A \code{flexseq} (left side).}

\item{y}{A \code{flexseq} (right side).}
}
\value{
Concatenated tree.
}
\description{
Same-name monoids are assumed equivalent; left-tree definitions win.
Missing monoids are added to each side before concatenation.
}
\keyword{internal}
