% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_trees.R
\name{concat_trees}
\alias{concat_trees}
\title{Concatenate two trees}
\usage{
concat_trees(x, y)
}
\arguments{
\item{x}{Left tree.}

\item{y}{Right tree.}
}
\value{
Concatenated tree.
}
\description{
Same-name monoids are assumed equivalent; left-tree definitions win.
Missing monoids are added to each side before concatenation.
}
