% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{concat_trees}
\alias{concat_trees}
\title{Concatenate two trees}
\usage{
concat_trees(x, y, monoid = NULL)
}
\arguments{
\item{x}{Left tree.}

\item{y}{Right tree.}

\item{monoid}{Optional monoid to store on the result.}
}
\value{
Concatenated tree.
}
\description{
Concatenate two trees
}
