% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-concat-trees.R
\name{concat_trees}
\alias{concat_trees}
\title{Concatenate Two Structural Trees}
\usage{
concat_trees(x, y)
}
\arguments{
\item{x}{A `flexseq` (left side).}

\item{y}{A `flexseq` (right side).}
}
\value{
Concatenated tree.
}
\description{
Same-name monoids are assumed equivalent; left-tree definitions win.
Missing monoids are added to each side before concatenation.
}
\examples{
\dontrun{
left <- as_flexseq(letters[1:3])
right <- as_flexseq(letters[4:6])
t <- concat_trees(left, right)
cat_m <- measure_monoid(paste0, "", as.character)
fold_left(t, cat_m)
}
}
\keyword{internal}
