% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-docs.R
\name{concat_trees}
\alias{concat_trees}
\title{Concatenate two trees}
\usage{
concat_trees(x, y, reducer = NULL)
}
\arguments{
\item{x}{Left tree.}

\item{y}{Right tree.}

\item{reducer}{Optional reducer to store on the tree.}
}
\description{
Concatenate two trees
}
