% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-ordered_sequence-ops.R
\name{count_key}
\alias{count_key}
\title{Count Elements Matching One Key}
\usage{
count_key(x, key)
}
\arguments{
\item{x}{An `ordered_sequence`.}

\item{key}{Query key.}
}
\value{
Integer count of matches.
}
\description{
Count Elements Matching One Key
}
\examples{
x <- ordered_sequence("a", "b", "c", keys = c(1, 2, 2))
count_key(x, 2)
count_key(x, 10)
}
