% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-ordered_sequence-ops.R
\name{count_key}
\alias{count_key}
\title{Count elements matching one key}
\usage{
count_key(x, key)
}
\arguments{
\item{x}{An `ordered_sequence`.}

\item{key}{Query key.}
}
\value{
Integer count of matching elements.
}
\description{
Count elements matching one key
}
