% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_multiset-core.R
\name{count_key}
\alias{count_key}
\title{Count multiplicity of one key}
\usage{
count_key(x, key_value)
}
\arguments{
\item{x}{An `ordered_multiset`.}

\item{key_value}{Query key.}
}
\value{
Integer count.
}
\description{
Count multiplicity of one key
}
