% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_sequence-core.R
\name{delete_all}
\alias{delete_all}
\title{Delete all elements by key}
\usage{
delete_all(x, key_value)
}
\arguments{
\item{x}{An `ordered_sequence`.}

\item{key_value}{Key value to delete.}
}
\value{
Updated `ordered_sequence`.
}
\description{
Delete all elements by key
}
