% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_sequence-core.R
\name{elements_between}
\alias{elements_between}
\title{Return elements in a key range}
\usage{
elements_between(x, lo, hi, include_lo = TRUE, include_hi = TRUE)
}
\arguments{
\item{x}{An `ordered_sequence`.}

\item{lo}{Lower bound key.}

\item{hi}{Upper bound key.}

\item{include_lo}{Include lower bound when `TRUE`.}

\item{include_hi}{Include upper bound when `TRUE`.}
}
\value{
List of raw elements.
}
\description{
Return elements in a key range
}
