% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{empty_tree}
\alias{empty_tree}
\title{Create an empty finger tree}
\usage{
empty_tree(reducer = NULL)
}
\arguments{
\item{reducer}{Optional reducer stored on the tree for later use.}
}
\value{
An empty finger tree.
}
\description{
Create an empty finger tree
}
\examples{
r <- Reducer(function(a, b) a + b, 0)
t <- empty_tree(reducer = r)
}
