% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_tree.R
\name{empty_tree}
\alias{empty_tree}
\title{Create an empty finger tree}
\usage{
empty_tree(monoids = NULL)
}
\arguments{
\item{monoids}{Optional named list of `MeasureMonoid` objects.}
}
\value{
An empty finger tree with structural `monoids` and `measures` attrs.
}
\description{
Create an empty finger tree
}
\examples{
t <- empty_tree()
t

count_m <- MeasureMonoid(`+`, 0, function(el) 1)
t2 <- empty_tree(monoids = list(count = count_m))
attr(t2, "measures")
}
