% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-empty-tree.R
\name{empty_tree}
\alias{empty_tree}
\title{Create an Empty Structural Tree}
\usage{
empty_tree(monoids = NULL)
}
\arguments{
\item{monoids}{Optional named list of `measure_monoid` objects.}
}
\value{
An empty finger tree with structural `monoids` and `measures` attrs.
}
\description{
Create an Empty Structural Tree
}
\examples{
\dontrun{
t <- empty_tree()
t

count_m <- measure_monoid(`+`, 0, function(el) 1)
t2 <- empty_tree(monoids = list(count = count_m))
attr(t2, "measures")
}
}
\keyword{internal}
