% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-seq-apply.R
\name{fapply.flexseq}
\alias{fapply.flexseq}
\title{Apply a function over flexseq elements}
\usage{
\method{fapply}{flexseq}(X, FUN, ...)
}
\arguments{
\item{X}{A `flexseq`.}

\item{FUN}{Function to apply to each element.}

\item{...}{Additional arguments passed to `FUN`.}
}
\value{
A new `flexseq` with transformed elements.
}
\description{
Apply a function over flexseq elements
}
