% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-apply-generics.R
\name{fapply}
\alias{fapply}
\title{Fapply with S3 dispatch}
\usage{
fapply(X, FUN, ...)
}
\arguments{
\item{X}{Object to apply over.}

\item{FUN}{Function to apply.}

\item{...}{Method-specific arguments.}
}
\value{
Method-dependent result.
}
\description{
`fapply()` is an S3 generic for applying functions over immutable
structures with type-specific dispatch.
}
\seealso{
[fapply.flexseq()], [fapply.priority_queue()], [fapply.ordered_sequence()]
}
