% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-core.R
\name{flexseq}
\alias{flexseq}
\title{Construct a Persistent Flexible Sequence}
\usage{
flexseq(...)
}
\arguments{
\item{...}{Sequence elements.}
}
\value{
A `flexseq` object.
}
\description{
Works like `list(...)`, but returns an immutable sequence backed by
measured finger-tree internals.
}
\examples{
x <- flexseq(1, 2, 3)
x

x2 <- flexseq("a", "b", "c")
x2
}
