% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexseq.R
\name{flexseq}
\alias{flexseq}
\title{Construct a Persistent Flexible Sequence}
\usage{
flexseq(..., monoids = NULL)
}
\arguments{
\item{...}{Sequence elements.}

\item{monoids}{Optional named list of measure monoids.}
}
\value{
A `flexseq` object.
}
\description{
Works like `list(...)`, but returns an immutable sequence backed by
measured finger-tree internals.
}
\examples{
x <- flexseq(1, 2, 3)
x[[2]]

y <- flexseq(a = "x", b = "y")
y$a
}
