% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexseq.R
\name{fold_right}
\alias{fold_right}
\title{Fold Right Over a Sequence}
\usage{
fold_right(x, monoid)
}
\arguments{
\item{x}{A `flexseq`.}

\item{monoid}{A `measure_monoid` specification.}
}
\value{
Reduced value.
}
\description{
Fold Right Over a Sequence
}
