% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_right.R
\name{fold_right}
\alias{fold_right}
\title{Fold Right Over a Sequence}
\usage{
fold_right(x, monoid)
}
\arguments{
\item{x}{A `flexseq`.}

\item{monoid}{A `measure_monoid` specification.}
}
\value{
Reduced value.
}
\description{
Fold Right Over a Sequence
}
\examples{
x <- as_flexseq(1:5)
x

sum_m <- measure_monoid(`+`, 0, as.numeric)
fold_right(x, sum_m)

x2 <- as_flexseq(letters[1:4])
cat_m <- measure_monoid(paste0, "", as.character)
fold_right(x2, cat_m)
}
