% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_graph_df}
\alias{get_graph_df}
\title{Build graph data frames for a finger tree}
\usage{
get_graph_df(t)
}
\arguments{
\item{t}{FingerTree.}
}
\value{
A list with edge and node data frames for igraph plotting.
}
\description{
Build graph data frames for a finger tree
}
\examples{
t <- tree_from(letters[1:4])
gdf <- get_graph_df(t)
names(gdf)
nrow(gdf[[1]])

# Works for deeper trees too
t2 <- tree_from(letters[1:10])
gdf2 <- get_graph_df(t2)
nrow(gdf2[[2]])
}
