% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immutables-package.R
\docType{package}
\name{immutables-package}
\alias{immutables}
\alias{immutables-package}
\title{immutables}
\description{
Sequence objects (\code{flexseq()}) supporting indexed and named access,
appending and prepending, concatenation, splitting, fast push
and pop from either end, item removal, and more.
}
\details{
Also implemented are priority queues (\code{priority_queue()})
focused on queue-oriented operations, including min and max
peeking and popping by priority value.

Backed by monoid-annotated 2-3 fingertrees, all structures are
persistent (operations return effective modified copies), and
most operations are constant time, amortized constant time, or
\eqn{O(\log n)} (indexing k elements is \eqn{O(k \log n)}). The
developer API supports the addition of custom structures
via combinations of monoids and measures; see vignettes for
details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oneilsh.github.io/immutables}
  \item \url{https://github.com/oneilsh/immutables}
  \item Report bugs at \url{https://github.com/oneilsh/immutables/issues}
}

}
\author{
\strong{Maintainer}: Shawn T. O'Neil \email{shawn@tislab.org}

}
\keyword{internal}
