% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingertree-package.R
\docType{package}
\name{immutables-package}
\alias{immutables}
\alias{immutables-package}
\title{immutables}
\description{
Monoid-annotated 2-3 finger trees.
}
\author{
\strong{Maintainer}: Shawn T. O'Neil \email{shawn@tislab.org}

}
\keyword{internal}
