% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-interval_index-core.R
\name{insert.interval_index}
\alias{insert.interval_index}
\title{Insert an element into an interval index}
\usage{
\method{insert}{interval_index}(x, element, start, end, name = NULL, ...)
}
\arguments{
\item{x}{An `interval_index`.}

\item{element}{Element to insert.}

\item{start}{Scalar start endpoint.}

\item{end}{Scalar end endpoint.}

\item{name}{Optional element name.}

\item{...}{Unused.}
}
\value{
Updated `interval_index`.
}
\description{
Insert an element into an interval index
}
