% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-ordered_sequence-ops.R
\name{insert.ordered_sequence}
\alias{insert.ordered_sequence}
\title{Insert an element into an ordered sequence}
\usage{
\method{insert}{ordered_sequence}(x, element, key, ...)
}
\arguments{
\item{x}{An `ordered_sequence`.}

\item{element}{Element to insert.}

\item{key}{Scalar key for `element`.}

\item{...}{Unused.}
}
\value{
Updated `ordered_sequence`.
}
\description{
Insert an element into an ordered sequence
}
