% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-queue-ops.R
\name{insert.priority_queue}
\alias{insert.priority_queue}
\title{Insert an element into a priority queue}
\usage{
\method{insert}{priority_queue}(x, element, priority, name = NULL, ...)
}
\arguments{
\item{x}{A `priority_queue`.}

\item{element}{Element to insert.}

\item{priority}{Scalar non-missing orderable priority.}

\item{name}{Optional element name.}

\item{...}{Unused.}
}
\value{
Updated `priority_queue`.
}
\description{
Insert an element into a priority queue
}
\examples{
x <- priority_queue("a", "b", priorities = c(2, 1))
x

x2 <- insert(x, "c", priority = 1)
x2
peek_min(x2)
}
