% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priority_queue.R
\name{insert}
\alias{insert}
\title{Insert an element into a priority queue}
\usage{
insert(q, element, priority, name = NULL)
}
\arguments{
\item{q}{A `priority_queue`.}

\item{element}{Element to insert.}

\item{priority}{Numeric scalar priority.}

\item{name}{Optional element name.}
}
\value{
Updated `priority_queue`.
}
\description{
Insert an element into a priority queue
}
\examples{
x <- priority_queue("a", "b", priorities = c(2, 1))
x

x2 <- insert(x, "c", priority = 1)
x2
peek_min(x2)
}
