% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_sequence-core.R,
%   R/40-priority_queue-core.R
\name{insert.ordered_sequence}
\alias{insert.ordered_sequence}
\alias{insert}
\alias{insert.priority_queue}
\title{Insert an element}
\usage{
\method{insert}{ordered_sequence}(x, element, key, ...)

insert(x, ...)

\method{insert}{priority_queue}(x, element, priority, name = NULL, ...)
}
\arguments{
\item{x}{Object to insert into.}

\item{element}{Element to insert.}

\item{key}{Scalar key for `element`.}

\item{...}{Method-specific arguments.}

\item{priority}{Numeric scalar priority.}

\item{name}{Optional element name.}
}
\value{
Updated `ordered_sequence`.

Updated object.
}
\description{
Insert an element
}
\details{
Generic `insert()` dispatches by class.
}
\examples{
x <- priority_queue("a", "b", priorities = c(2, 1))
x

x2 <- insert(x, "c", priority = 1)
x2
peek_min(x2)
}
