% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-core.R
\name{insert}
\alias{insert}
\title{Insert an element}
\usage{
insert(x, ...)
}
\arguments{
\item{x}{Object to insert into.}

\item{...}{Method-specific arguments.}
}
\value{
Updated object.
}
\description{
Generic `insert()` dispatches by class.
}
\seealso{
[insert.priority_queue()], [insert.ordered_sequence()], [insert.interval_index()]
}
