% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-query-api.R
\name{interval_bounds}
\alias{interval_bounds}
\title{Get interval bounds in sequence order}
\usage{
interval_bounds(x)
}
\arguments{
\item{x}{An `interval_index`.}
}
\value{
Data frame with list-columns `start` and `end`.
}
\description{
Get interval bounds in sequence order
}
