% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-interval_index-core.R
\name{interval_index}
\alias{interval_index}
\title{Construct an Interval Index}
\usage{
interval_index(..., start = NULL, end = NULL, bounds = "[)", monoids = NULL)
}
\arguments{
\item{...}{Elements to add.}

\item{start}{Scalar start endpoints matching `...`.}

\item{end}{Scalar end endpoints matching `...`.}

\item{bounds}{One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}

\item{monoids}{Optional additional named list of `measure_monoid` objects.}
}
\value{
An `interval_index`.
}
\description{
Construct an Interval Index
}
\examples{
ix <- interval_index("a", "b", "c", start = c(1, 2, 2), end = c(3, 2, 4))
ix
}
