% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-constructors.R
\name{interval_index}
\alias{interval_index}
\title{Construct an Interval Index}
\usage{
interval_index(..., start, end, bounds = "[)")
}
\arguments{
\item{...}{Elements to add.}

\item{start}{Scalar start endpoints matching `...`.}

\item{end}{Scalar end endpoints matching `...`.}

\item{bounds}{One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}
}
\value{
An `interval_index`.
}
\description{
Construct an Interval Index
}
\examples{
ix <- interval_index("a", "b", "c", start = c(1, 2, 2), end = c(3, 2, 4))
ix
}
