% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-core.R,
%   R/30-api-is-empty-generic.R
\name{is_empty.flexseq}
\alias{is_empty.flexseq}
\alias{is_empty}
\title{Check emptiness with S3 dispatch}
\usage{
\method{is_empty}{flexseq}(x, ...)

is_empty(x, ...)
}
\arguments{
\item{x}{Object to check.}

\item{...}{Reserved for method-specific arguments.}
}
\value{
Logical scalar.
}
\description{
Generic `is_empty()` dispatches by class.
}
\examples{
is_empty(flexseq())
is_empty(as_flexseq(1:3))

is_empty(priority_queue())

ms <- as_ordered_multiset(list("a"), keys = 1)
is_empty(ms)
}
