% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-core.R
\name{length.flexseq}
\alias{length.flexseq}
\title{Sequence Length}
\usage{
\method{length}{flexseq}(x)
}
\arguments{
\item{x}{A `flexseq`.}
}
\value{
Number of elements in the sequence.
}
\description{
Sequence Length
}
