% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-core.R,
%   R/40-priority_queue-core.R
\name{length.flexseq}
\alias{length.flexseq}
\alias{length.priority_queue}
\title{Sequence Length}
\usage{
\method{length}{flexseq}(x)

\method{length}{priority_queue}(x)
}
\arguments{
\item{x}{A `flexseq`.}
}
\value{
Number of elements in the sequence.
}
\description{
Sequence Length
}
