% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-methods.R
\name{length.interval_index}
\alias{length.interval_index}
\title{Interval Index Length}
\usage{
\method{length}{interval_index}(x)
}
\arguments{
\item{x}{An `interval_index`.}
}
\value{
Number of indexed intervals.
}
\description{
Interval Index Length
}
\examples{
ix <- interval_index("a", "b", start = c(1, 3), end = c(2, 5))
length(ix)

length(interval_index())
}
