% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-interval_index-core.R
\name{length.interval_index}
\alias{length.interval_index}
\title{Interval Index Length}
\usage{
\method{length}{interval_index}(x)
}
\arguments{
\item{x}{An `interval_index`.}
}
\value{
Integer length.
}
\description{
Interval Index Length
}
