% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_sequence-core.R
\name{length.ordered_sequence}
\alias{length.ordered_sequence}
\title{Ordered Sequence Length}
\usage{
\method{length}{ordered_sequence}(x)
}
\arguments{
\item{x}{An `ordered_sequence`.}
}
\value{
Integer length.
}
\description{
Ordered Sequence Length
}
