% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-methods.R
\name{length.priority_queue}
\alias{length.priority_queue}
\title{Priority Queue Length}
\usage{
\method{length}{priority_queue}(x)
}
\arguments{
\item{x}{A `priority_queue`.}
}
\value{
Integer length.
}
\description{
Priority Queue Length
}
