% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureMonoid.R
\name{measure_monoid}
\alias{measure_monoid}
\title{Construct a Measure Monoid Specification}
\usage{
measure_monoid(f, i, measure)
}
\arguments{
\item{f}{Associative binary function over measure values.}

\item{i}{Identity measure value for `f`.}

\item{measure}{Function mapping a raw element to a measure value.}
}
\value{
An object of class `measure_monoid`.
}
\description{
Construct a Measure Monoid Specification
}
\examples{
sum_m <- measure_monoid(`+`, 0, as.numeric)
x <- as_flexseq(1:5)
fold_left(x, sum_m)

# create a second monoid and add it to a sequence
nchar_sum <- measure_monoid(`+`, 0, function(el) nchar(as.character(el)))
x2 <- as_flexseq(letters[1:3])
x3 <- add_monoids(x2, list(nchar_sum = nchar_sum))
attr(x3, "measures")$nchar_sum
}
