% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureMonoid.R
\name{measure_monoid}
\alias{measure_monoid}
\title{Construct a Measure Monoid Specification}
\usage{
measure_monoid(f, i, measure)
}
\arguments{
\item{f}{Associative binary function over measure values.}

\item{i}{Identity measure value for `f`.}

\item{measure}{Function mapping a raw element to a measure value.}
}
\value{
An object of class `measure_monoid`.
}
\description{
Construct a Measure Monoid Specification
}
\examples{
sum_m <- measure_monoid(`+`, 0, as.numeric)
t <- as_flexseq(1:5)
fold_left(t, sum_m)

# Measure from element metadata
t2 <- as_flexseq(letters[1:3], values = c(10, 20, 30))
value_sum <- measure_monoid(`+`, 0, function(el) as.numeric(attr(el, "value")))
t2m <- add_monoids(t2, list(value_sum = value_sum))
attr(t2m, "measures")$value_sum
}
