% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-docs.R
\name{MeasureMonoid}
\alias{MeasureMonoid}
\title{Measured monoid}
\usage{
MeasureMonoid(...)
}
\arguments{
\item{f}{Reduction function.}

\item{i}{Identity element.}

\item{measure}{Function mapping an element to a measure.}
}
\description{
Measured monoid
}
