% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureMonoid.R
\name{MeasureMonoid}
\alias{MeasureMonoid}
\title{Construct a MeasureMonoid}
\usage{
MeasureMonoid(f, i, measure)
}
\arguments{
\item{f}{Associative binary function over measure values.}

\item{i}{Identity measure value for `f`.}

\item{measure}{Function mapping a raw element to a measure value.}
}
\value{
An object of class `MeasureMonoid`.
}
\description{
Construct a MeasureMonoid
}
\examples{
sum_m <- MeasureMonoid(`+`, 0, as.numeric)
t <- tree_from(1:5)
reduce_left(t, sum_m)

# Measure from element metadata
t2 <- tree_from(letters[1:3], values = c(10, 20, 30))
value_sum <- MeasureMonoid(`+`, 0, function(el) as.numeric(attr(el, "value")))
reduce_right(t2, value_sum)
}
