% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_multiset-core.R
\name{ordered_multiset}
\alias{ordered_multiset}
\title{Construct an Ordered Multiset}
\usage{
ordered_multiset(..., keys = NULL, monoids = NULL)
}
\arguments{
\item{...}{Elements to add.}

\item{keys}{Scalar key values matching `...` length.}

\item{monoids}{Optional additional named list of `measure_monoid` objects.}
}
\value{
An `ordered_multiset`.
}
\description{
Construct an Ordered Multiset
}
\examples{
ms <- ordered_multiset("bb", "a", "ccc", keys = c(2, 1, 3))
ms
union(ms, ms)
}
