% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_sequence-core.R
\name{ordered_sequence}
\alias{ordered_sequence}
\title{Construct an Ordered Sequence}
\usage{
ordered_sequence(..., keys = NULL, monoids = NULL)
}
\arguments{
\item{...}{Elements to add.}

\item{keys}{Scalar key values matching `...` length.}

\item{monoids}{Optional additional named list of `measure_monoid` objects.}
}
\value{
An `ordered_sequence`.
}
\description{
Construct an Ordered Sequence
}
\examples{
xs <- ordered_sequence("bb", "a", "ccc", keys = c(2, 1, 3))
xs
lower_bound(xs, 2)
}
