% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-ordered_sequence-ops.R
\name{peek_all_key}
\alias{peek_all_key}
\title{Peek All Elements for One Key}
\usage{
peek_all_key(x, key)
}
\arguments{
\item{x}{An \code{ordered_sequence}.}

\item{key}{Query key.}
}
\value{
An \code{ordered_sequence} containing all matches; empty on miss.
}
\description{
Returns all elements whose key equals \code{key}.
}
\details{
The returned \code{ordered_sequence} can be inspected with \code{\link[=as.list]{as.list()}}.
}
\examples{
x <- ordered_sequence("a", "b", "c", keys = c(1, 2, 2))
out <- peek_all_key(x, 2)
as.list(out)
}
