% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-query-api.R
\name{peek_all_point}
\alias{peek_all_point}
\title{Peek all intervals containing a point}
\usage{
peek_all_point(x, point, bounds = NULL)
}
\arguments{
\item{x}{An `interval_index`.}

\item{point}{Query point.}

\item{bounds}{Optional boundary override. One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}
}
\value{
An `interval_index` slice of all matches (possibly empty).
}
\description{
Peek all intervals containing a point
}
