% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-query-api.R
\name{peek_all_point}
\alias{peek_all_point}
\title{Peek All Intervals Containing a Point}
\usage{
peek_all_point(x, point, bounds = NULL)
}
\arguments{
\item{x}{An \code{interval_index}.}

\item{point}{Query point.}

\item{bounds}{Optional boundary override. One of \code{"[)"}, \code{"[]"}, \code{"()"}, \code{"(]"}.}
}
\value{
An \code{interval_index} slice of all matches (possibly empty).
}
\description{
Peek All Intervals Containing a Point
}
\details{
The returned \code{interval_index} can be inspected with \code{\link[=as.list]{as.list()}}.
}
\examples{
ix <- interval_index("a", "b", "c", start = c(1, 2, 4), end = c(3, 2, 5))
as.list(peek_all_point(ix, 2))
}
