% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-query-api.R
\name{peek_all_within}
\alias{peek_all_within}
\title{Peek All Intervals Within a Query Interval}
\usage{
peek_all_within(x, start, end, bounds = NULL)
}
\arguments{
\item{x}{An \code{interval_index}.}

\item{start}{Query interval start.}

\item{end}{Query interval end.}

\item{bounds}{Optional boundary override. One of \code{"[)"}, \code{"[]"}, \code{"()"}, \code{"(]"}.}
}
\value{
An \code{interval_index} slice of all matches (possibly empty).
}
\description{
Peek All Intervals Within a Query Interval
}
\details{
The returned \code{interval_index} can be inspected with \code{\link[=as.list]{as.list()}}.
}
\examples{
ix <- interval_index("a", "b", "c", start = c(1, 2, 4), end = c(6, 3, 5))
as.list(peek_all_within(ix, 1, 4))
}
