% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-ends.R
\name{peek_at}
\alias{peek_at}
\title{Peek at an element by position}
\usage{
peek_at(x, index)
}
\arguments{
\item{x}{A `flexseq`.}

\item{index}{One-based position to read.}
}
\value{
Element at `index`.
}
\description{
Peek at an element by position
}
