% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-ends.R
\name{peek_back}
\alias{peek_back}
\title{Peek at the back element}
\usage{
peek_back(x)
}
\arguments{
\item{x}{A `flexseq`.}
}
\value{
Back element, or `NULL` when `x` is empty.
}
\description{
Peek at the back element
}
