% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-ends.R
\name{peek_front}
\alias{peek_front}
\title{Peek at the front element}
\usage{
peek_front(x)
}
\arguments{
\item{x}{A `flexseq`.}
}
\value{
Front element, or `NULL` when `x` is empty.
}
\description{
Peek at the front element
}
