% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-queue-ops.R
\name{peek_min}
\alias{peek_min}
\title{Peek minimum-priority element}
\usage{
peek_min(x)
}
\arguments{
\item{x}{A `priority_queue`.}
}
\value{
Element with minimum priority (stable on ties), or `NULL` when empty.
}
\description{
Peek minimum-priority element
}
\examples{
x <- priority_queue("a", "b", "c", priorities = c(2, 1, 1))
x
peek_min(x)
}
