% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priority_queue.R
\name{peek_min}
\alias{peek_min}
\title{Peek minimum-priority element}
\usage{
peek_min(q)
}
\arguments{
\item{q}{A `priority_queue`.}
}
\value{
Element with minimum priority (stable on ties).
}
\description{
Peek minimum-priority element
}
\examples{
q <- priority_queue("a", "b", "c", priorities = c(2, 1, 1))
peek_min(q)
}
