% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-query-api.R
\name{peek_point}
\alias{peek_point}
\title{Peek first interval containing a point}
\usage{
peek_point(x, point, bounds = NULL)
}
\arguments{
\item{x}{An `interval_index`.}

\item{point}{Query point.}

\item{bounds}{Optional boundary override. One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}
}
\value{
The payload item from the first match, or `NULL` on no match.
}
\description{
Peek first interval containing a point
}
