% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-core.R
\name{plot.flexseq}
\alias{plot.flexseq}
\title{Plot a Sequence Tree}
\usage{
\method{plot}{flexseq}(x, ...)
}
\arguments{
\item{x}{A \code{flexseq}.}

\item{...}{Passed to the internal tree plotting routine.}
}
\description{
Plot a Sequence Tree
}
\details{
Visualizes the internal finger-tree structure, not a value-level chart.
}
\examples{
x <- flexseq("a", "b", "c")
plot(x)
}
