% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-methods.R
\name{plot.interval_index}
\alias{plot.interval_index}
\title{Plot an Interval Index Tree}
\usage{
\method{plot}{interval_index}(x, ...)
}
\arguments{
\item{x}{An `interval_index`.}

\item{...}{Passed to the internal tree plotting routine.}
}
\description{
Plot an Interval Index Tree
}
