% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-core.R
\name{plot.ordered_sequence}
\alias{plot.ordered_sequence}
\title{Plot an Ordered Sequence Tree}
\usage{
\method{plot}{ordered_sequence}(x, ...)
}
\arguments{
\item{x}{An `ordered_sequence`.}

\item{...}{Passed to the internal tree plotting routine.}
}
\description{
Plot an Ordered Sequence Tree
}
