% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-methods.R
\name{plot.priority_queue}
\alias{plot.priority_queue}
\title{Plot a Priority Queue Tree}
\usage{
\method{plot}{priority_queue}(x, ...)
}
\arguments{
\item{x}{A `priority_queue`.}

\item{...}{Passed to the internal tree plotting routine.}
}
\description{
Plot a Priority Queue Tree
}
