% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-docs.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a finger tree}
\usage{
plot_tree(
  t1,
  vertex.size = 4,
  edge.width = 1,
  label_edges = FALSE,
  title = NULL,
  ...
)
}
\arguments{
\item{t}{FingerTree.}
}
\description{
Plot a finger tree
}
