% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a finger tree with igraph}
\usage{
plot_tree(
  t1,
  vertex.size = 4,
  edge.width = 1,
  label_edges = FALSE,
  title = NULL,
  node_label = c("value", "type", "both", "none"),
  ...
)
}
\arguments{
\item{t1}{FingerTree.}

\item{vertex.size}{Vertex size passed to `igraph::plot.igraph`.}

\item{edge.width}{Edge width passed to `igraph::plot.igraph`.}

\item{label_edges}{Whether to draw edge labels.}

\item{title}{Optional plot title.}

\item{node_label}{Node label mode: value, type, both, or none.}

\item{...}{Additional arguments passed to `igraph::plot.igraph`.}
}
\description{
Plot a finger tree with igraph
}
\examples{
\dontrun{
t <- tree_from(letters[1:8])
plot_tree(t, title = "Finger tree")

t2 <- tree_from(letters[1:12])
plot_tree(t2, node_label = "both", vertex.size = 8)
}
}
