% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-query-api.R
\name{pop_all_containing}
\alias{pop_all_containing}
\title{Pop All Containing Intervals}
\usage{
pop_all_containing(x, start, end, bounds = NULL)
}
\arguments{
\item{x}{An `interval_index`.}

\item{start}{Query interval start.}

\item{end}{Query interval end.}

\item{bounds}{Optional boundary override. One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}
}
\value{
A list with `elements` and `remaining`, both `interval_index` objects.
}
\description{
Pop All Containing Intervals
}
\details{
Use [as.list()] to convert `elements` to a standard R list.
}
\examples{
ix <- interval_index("a", "b", "c", start = c(1, 2, 4), end = c(6, 5, 7))
out <- pop_all_containing(ix, 2, 4)
as.list(out$elements)
}
