% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-queue-ops.R
\name{pop_all_max}
\alias{pop_all_max}
\title{Pop all maximum-priority elements}
\usage{
pop_all_max(x)
}
\arguments{
\item{x}{A `priority_queue`.}
}
\value{
List with `elements` and updated `remaining`.
}
\description{
Pop all maximum-priority elements
}
