% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60-interval_index-query-api.R
\name{pop_all_overlaps}
\alias{pop_all_overlaps}
\title{Pop All Overlapping Intervals}
\usage{
pop_all_overlaps(x, start, end, bounds = NULL)
}
\arguments{
\item{x}{An `interval_index`.}

\item{start}{Query interval start.}

\item{end}{Query interval end.}

\item{bounds}{Optional boundary override. One of `"[)"`, `"[]"`, `"()"`, `"(]"`.}
}
\value{
A list with `elements` and `remaining`, both `interval_index` objects.
}
\description{
Pop All Overlapping Intervals
}
\details{
Use [as.list()] to convert `elements` to a standard R list.
}
\examples{
ix <- interval_index("a", "b", "c", start = c(1, 3, 5), end = c(2, 4, 6))
out <- pop_all_overlaps(ix, 2, 5)
as.list(out$elements)
}
