% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-ends.R
\name{pop_back}
\alias{pop_back}
\title{Pop the back element}
\usage{
pop_back(x)
}
\arguments{
\item{x}{A `flexseq`.}
}
\value{
A list with fields `element` and `rest`.
}
\description{
Returns both the popped element and the remaining sequence.
}
\examples{
s <- as_flexseq(letters[1:3])
out <- pop_back(s)
out$element
out$rest
}
