% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-ends.R
\name{pop_front}
\alias{pop_front}
\title{Pop the front element}
\usage{
pop_front(x)
}
\arguments{
\item{x}{A `flexseq`.}
}
\value{
A list with fields `value` and `rest`.
}
\description{
Returns both the popped value and the remaining sequence.
}
\examples{
s <- as_flexseq(letters[1:3])
out <- pop_front(s)
out$value
out$rest
}
