% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-queue-ops.R
\name{pop_max}
\alias{pop_max}
\title{Pop maximum-priority element}
\usage{
pop_max(x)
}
\arguments{
\item{x}{A `priority_queue`.}
}
\value{
List with `element`, `priority`, and updated `remaining`.
}
\description{
Pop maximum-priority element
}
\examples{
x <- priority_queue("a", "b", "c", priorities = c(2, 3, 3))
out <- pop_max(x)
out$element
out$priority
out$remaining
}
