% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-priority_queue-queue-ops.R
\name{pop_min}
\alias{pop_min}
\title{Pop minimum-priority element}
\usage{
pop_min(x)
}
\arguments{
\item{x}{A `priority_queue`.}
}
\value{
List with `element`, `priority`, and updated `remaining`.
}
\description{
Pop minimum-priority element
}
\examples{
x <- priority_queue("a", "b", "c", priorities = c(2, 1, 1))
out <- pop_min(x)
out$element
out$priority
out$remaining
}
