% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predicate.R
\name{Predicate}
\alias{Predicate}
\title{Construct a split predicate wrapper}
\usage{
Predicate(f)
}
\arguments{
\item{f}{Predicate function over accumulated measure values.}
}
\value{
The predicate function.
}
\description{
Construct a split predicate wrapper
}
