% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predicate.R
\name{predicate}
\alias{predicate}
\title{Construct a Predicate Function}
\usage{
predicate(f)
}
\arguments{
\item{f}{Predicate function over accumulated measure values.}
}
\value{
The predicate function.
}
\description{
Construct a Predicate Function
}
\examples{
p <- predicate(function(v) v >= 3)
p(2)
p(3)

x <- as_flexseq(letters[1:5])
locate(x, p, ".size")
}
