% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{prepend}
\alias{prepend}
\title{Prepend an element}
\usage{
prepend(t, x, monoid = NULL)
}
\arguments{
\item{t}{FingerTree.}

\item{x}{Element to prepend.}

\item{monoid}{Optional monoid to store on the tree.}
}
\value{
Updated tree.
}
\description{
Prepend an element
}
