% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-prepend.R
\name{prepend}
\alias{prepend}
\title{Prepend an element}
\usage{
prepend(t, x)
}
\arguments{
\item{t}{FingerTree.}

\item{x}{Element to prepend.}
}
\value{
Updated tree.
  If `t` is name-indexed, prepending unnamed elements is invalid.
}
\description{
Prepend an element
}
\examples{
x <- as_flexseq(letters[2:5])
x

x2 <- prepend(x, "a")
x2

# prepending to a named sequence requires a named element
x3 <- as_flexseq(list(two = 2, three = 3))
new_el <- 1
names(new_el) <- "one"
x4 <- prepend(x3, new_el)
x4
}
