% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FingerTree.R
\name{print.FingerTree}
\alias{print.FingerTree}
\alias{print.Deep}
\alias{print.Single}
\alias{print.Empty}
\title{Print a compact summary of a finger tree}
\usage{
\method{print}{FingerTree}(x, max_elements = 6L, show_internal_monoids = FALSE, ...)

\method{print}{Deep}(x, ...)

\method{print}{Single}(x, ...)

\method{print}{Empty}(x, ...)
}
\arguments{
\item{x}{FingerTree.}

\item{max_elements}{Maximum number of elements to show in list-style preview.
Default `6`.}

\item{show_internal_monoids}{Logical; show internal monoids (`.size`,
`.named_count`). Default `FALSE`.}

\item{...}{Passed through to `print()` for preview elements.}
}
\value{
`x`, invisibly.
}
\description{
Print a compact summary of a finger tree
}
\examples{
x <- as_flexseq(letters[1:10])
x

x2 <- as_flexseq(setNames(as.list(1:5), paste0("k", 1:5)))
x2
}
\keyword{internal}
