% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-print.R
\name{print.flexseq}
\alias{print.flexseq}
\title{Print a flexseq}
\usage{
\method{print}{flexseq}(x, ...)
}
\arguments{
\item{x}{A `flexseq`.}

\item{...}{Passed through to `print.FingerTree`.}
}
\description{
Print a flexseq
}
