% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.priority_queue.R
\name{print.priority_queue}
\alias{print.priority_queue}
\title{Print a Priority Queue}
\usage{
\method{print}{priority_queue}(x, ...)
}
\arguments{
\item{x}{A `priority_queue`.}

\item{...}{Unused.}
}
\value{
`x` invisibly.
}
\description{
Print a Priority Queue
}
\examples{
x <- priority_queue("a", "b", "c", priorities = c(2, 1, 3))
x
}
