% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priority_queue.R
\name{priority_queue}
\alias{priority_queue}
\title{Construct a Priority Queue}
\usage{
priority_queue(..., priorities = NULL, names = NULL, monoids = NULL)
}
\arguments{
\item{...}{Elements to enqueue.}

\item{priorities}{Numeric priorities matching `...`.}

\item{names}{Optional element names.}

\item{monoids}{Optional additional named list of `measure_monoid` objects.}
}
\value{
A `priority_queue`.
}
\description{
Construct a Priority Queue
}
\examples{
q <- priority_queue("a", "b", "c", priorities = c(2, 1, 2))
peek_min(q)
}
