% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-ends.R
\name{push_back}
\alias{push_back}
\title{Push an element to the back}
\usage{
push_back(x, value)
}
\arguments{
\item{x}{A `flexseq`.}

\item{value}{Element to push.}
}
\value{
Updated tree.
  If `x` is name-indexed, pushing unnamed elements is invalid.
}
\description{
Push an element to the back
}
\examples{
s <- as_flexseq(letters[1:3])
s2 <- push_back(s, "d")
s2

n <- as_flexseq(list(two = 2, three = 3))
new_el <- 4
names(new_el) <- "four"
push_back(n, new_el)
}
