% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-ends.R
\name{push_front}
\alias{push_front}
\title{Push an element to the front}
\usage{
push_front(x, value)
}
\arguments{
\item{x}{A `flexseq`.}

\item{value}{Element to push.}
}
\value{
Updated tree.
  If `x` is name-indexed, pushing unnamed elements is invalid.
}
\description{
Push an element to the front
}
\examples{
s <- as_flexseq(letters[2:4])
s2 <- push_front(s, "a")
s2

n <- as_flexseq(list(two = 2, three = 3))
new_el <- 1
names(new_el) <- "one"
push_front(n, new_el)
}
