% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_left.R
\name{reduce_left}
\alias{reduce_left}
\title{Reduce from the left}
\usage{
reduce_left(t, monoid)
}
\arguments{
\item{t}{FingerTree.}

\item{monoid}{A `MeasureMonoid` object.}
}
\value{
Reduced value.
}
\description{
Reduce from the left
}
