% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{reduce_left}
\alias{reduce_left}
\title{Reduce from the left}
\usage{
reduce_left(t, reducer = NULL)
}
\arguments{
\item{t}{FingerTree.}

\item{reducer}{Optional reducer; falls back to tree attribute.}
}
\value{
Reduced value.
}
\description{
Reduce from the left
}
