% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_left.R
\name{reduce_left}
\alias{reduce_left}
\title{Reduce from the left}
\usage{
reduce_left(t, monoid)
}
\arguments{
\item{t}{FingerTree.}

\item{monoid}{A `MeasureMonoid` object.}
}
\value{
Reduced value.
}
\description{
Reduce from the left
}
\examples{
t <- tree_from(1:5)
sum_m <- MeasureMonoid(`+`, 0, as.numeric)
reduce_left(t, sum_m)

cat_m <- MeasureMonoid(paste0, "", as.character)
reduce_left(tree_from(letters[1:4]), cat_m)
}
