% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-docs.R
\name{reduce_right}
\alias{reduce_right}
\title{Reduce from the right}
\usage{
reduce_right(t, reducer = NULL)
}
\arguments{
\item{t}{FingerTree.}

\item{reducer}{Optional reducer; falls back to tree attribute.}
}
\description{
Reduce from the right
}
