% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_right.R
\name{reduce_right}
\alias{reduce_right}
\title{Reduce from the right}
\usage{
reduce_right(t, monoid)
}
\arguments{
\item{t}{FingerTree.}

\item{monoid}{A `MeasureMonoid` object.}
}
\value{
Reduced value.
}
\description{
Reduce from the right
}
