% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{reduce_right}
\alias{reduce_right}
\title{Reduce from the right}
\usage{
reduce_right(t, monoid = NULL)
}
\arguments{
\item{t}{FingerTree.}

\item{monoid}{Optional monoid; falls back to tree attribute.}
}
\value{
Reduced value.
}
\description{
Reduce from the right
}
