% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\title{Split tree into left and right parts}
\usage{
split(t, predicate, monoid_name)
}
\arguments{
\item{t}{FingerTree.}

\item{predicate}{Function on measure values.}

\item{monoid_name}{Name of monoid from `attr(t, "monoids")`.}
}
\value{
A list with `left` and `right`.
}
\description{
Split tree into left and right parts
}
\examples{
t <- tree_from(letters[1:6])
s <- split(t, function(v) v >= 4, ".size")
attr(s$left, "measures")$.size
attr(s$right, "measures")$.size

cat_m <- MeasureMonoid(paste0, "", as.character)
fold_left(s$left, cat_m)
fold_left(s$right, cat_m)
}
