% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\title{Split tree into left and right parts}
\usage{
split(t, predicate, monoid_name)
}
\arguments{
\item{t}{FingerTree.}

\item{predicate}{Function on measure values.}

\item{monoid_name}{Name of monoid from `attr(t, "monoids")`.}
}
\value{
A list with `left` and `right`.
}
\description{
Split tree into left and right parts
}
