% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tree.R
\name{split_tree}
\alias{split_tree}
\title{Split tree around first predicate flip}
\usage{
split_tree(t, predicate, monoid_name, accumulator = NULL)
}
\arguments{
\item{t}{FingerTree.}

\item{predicate}{Function on measure values.}

\item{monoid_name}{Name of monoid from `attr(t, "monoids")`.}

\item{accumulator}{Optional starting measure (defaults to monoid identity).}
}
\value{
A list with `left`, `elem`, and `right`.
}
\description{
Split tree around first predicate flip
}
\examples{
x <- as_flexseq(letters[1:6])
x

s <- split_tree(x, function(v) v >= 4, ".size")
s$elem
s$left
s$right
}
