% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-str.R
\name{str.flexseq}
\alias{str.flexseq}
\title{Display Internal Structure of a flexseq}
\usage{
\method{str}{flexseq}(object, ...)
}
\arguments{
\item{object}{A `flexseq`.}

\item{...}{Passed to [utils::str()].}
}
\value{
`NULL`, invisibly.
}
\description{
Display Internal Structure of a flexseq
}
