% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{[[.FingerTree}
\alias{[[.FingerTree}
\title{Extract a single element from a finger tree by 1-based index}
\usage{
\method{[[}{FingerTree}(x, i, ...)
}
\arguments{
\item{x}{FingerTree.}

\item{i}{Positive scalar integer index.}

\item{...}{Unused.}
}
\value{
The extracted element.
}
\description{
Extract a single element from a finger tree by 1-based index
}
