% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{[[.FingerTree}
\alias{[[.FingerTree}
\title{Extract one element by position or unique name}
\usage{
\method{[[}{FingerTree}(x, i, ...)
}
\arguments{
\item{x}{FingerTree.}

\item{i}{Positive scalar integer index, or scalar character element name.}

\item{...}{Unused.}
}
\value{
The extracted element (internal name metadata is removed).
}
\description{
Extract one element by position or unique name
}
\examples{
t <- tree_from(letters[1:5])
t[[3]]

tn <- tree_from(setNames(as.list(letters[1:3]), c("a1", "a2", "a3")))
tn[["a2"]]
}
