% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-api-flexseq-indexing.R
\name{[[.flexseq}
\alias{[[.flexseq}
\title{Extract one element by position or unique name}
\usage{
\method{[[}{flexseq}(x, i, ...)
}
\arguments{
\item{x}{A `flexseq`.}

\item{i}{Positive scalar integer index, or scalar character element name.}

\item{...}{Unused.}
}
\value{
The extracted element (internal name metadata is removed).
}
\description{
Extract one element by position or unique name
}
\examples{
x <- as_flexseq(letters[1:5])
x[[3]]

x2 <- as_flexseq(setNames(as.list(letters[1:3]), c("a1", "a2", "a3")))
x2[["a2"]]
}
