% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{[[<-.FingerTree}
\alias{[[<-.FingerTree}
\title{Replace a single element by 1-based index}
\usage{
\method{[[}{FingerTree}(x, i) <- value
}
\arguments{
\item{x}{FingerTree.}

\item{i}{Positive scalar integer index.}

\item{value}{Replacement element.}
}
\value{
A new FingerTree with one element replaced.
}
\description{
Replace a single element by 1-based index
}
