% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{[[<-.FingerTree}
\alias{[[<-.FingerTree}
\title{Replace one element by position or unique name}
\usage{
\method{[[}{FingerTree}(x, i) <- value
}
\arguments{
\item{x}{FingerTree.}

\item{i}{Positive scalar integer index, or scalar character element name.}

\item{value}{Replacement element.}
}
\value{
A new FingerTree with one element replaced.
}
\description{
Replace one element by position or unique name
}
\examples{
t <- tree_from(letters[1:4])
t[[2]] <- "ZZ"
cat_m <- MeasureMonoid(paste0, "", as.character)
reduce_left(t, cat_m)

tn <- tree_from(setNames(as.list(1:3), c("x", "y", "z")))
tn[["y"]] <- 99

# Assigning NULL removes an element (by index or name)
t <- tree_from(letters[1:4])
t[[2]] <- NULL
tn <- tree_from(setNames(as.list(1:3), c("a", "b", "c")))
tn[["b"]] <- NULL
}
