% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-docs.R
\name{tree_from}
\alias{tree_from}
\title{Build a tree from a vector or list}
\usage{
tree_from(x, values = NULL, reducer = NULL)
}
\arguments{
\item{x}{Elements to insert.}

\item{values}{Optional values for each element.}

\item{reducer}{Optional reducer to store on the tree.}
}
\description{
Build a tree from a vector or list
}
