% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40-ordered_multiset-core.R
\name{upper_bound}
\alias{upper_bound}
\title{Find first element with key > value}
\usage{
upper_bound(x, key_value)
}
\arguments{
\item{x}{An `ordered_multiset`.}

\item{key_value}{Query key.}
}
\value{
`list(found, index, element, key, seq_id)`.
}
\description{
Find first element with key > value
}
